+++ packages/messaging-api-line/src/__tests__/LineClient-menu.spec.js
@@ -12,12 +12,13 @@ const CHANNEL_SECRET = 'so-secret';
 const createMock = ({ customAccessToken } = {}) => {
   const client = new LineClient(ACCESS_TOKEN, CHANNEL_SECRET);
   const mock = new MockAdapter(client.axios);
+  const dataMock = new MockAdapter(client.dataAxios);
   const headers = {
     Accept: 'application/json, text/plain, */*',
     'Content-Type': 'application/json',
     Authorization: `Bearer ${customAccessToken || ACCESS_TOKEN}`,
   };
-  return { client, mock, headers };
+  return { client, mock, dataMock, headers };
 };
 
 describe('Rich Menu', () => {
@@ -588,7 +589,7 @@ describe('Rich Menu', () => {
     it('should call api', async () => {
       expect.assertions(4);
 
-      const { client, mock, headers } = createMock();
+      const { client, dataMock, headers } = createMock();
 
       const reply = {};
 
@@ -602,9 +603,9 @@ describe('Rich Menu', () => {
         });
       });
 
-      mock.onPost().reply(config => {
+      dataMock.onPost().reply(config => {
         expect(config.url).toEqual(
-          'https://api.line.me/v2/bot/richmenu/1/content'
+          'https://api-data.line.me/v2/bot/richmenu/1/content'
         );
         expect(config.data).toEqual(buffer);
         expect(config.headers).toEqual({
@@ -622,7 +623,7 @@ describe('Rich Menu', () => {
     it('should work with custom access token', async () => {
       expect.assertions(4);
 
-      const { client, mock, headers } = createMock({
+      const { client, dataMock, headers } = createMock({
         customAccessToken: CUSTOM_ACCESS_TOKEN,
       });
 
@@ -638,9 +639,9 @@ describe('Rich Menu', () => {
         });
       });
 
-      mock.onPost().reply(config => {
+      dataMock.onPost().reply(config => {
         expect(config.url).toEqual(
-          'https://api.line.me/v2/bot/richmenu/1/content'
+          'https://api-data.line.me/v2/bot/richmenu/1/content'
         );
         expect(config.data).toEqual(buffer);
         expect(config.headers).toEqual({
@@ -660,13 +661,13 @@ describe('Rich Menu', () => {
     it('should throw error when ', async () => {
       expect.assertions(1);
 
-      const { client, mock, headers } = createMock();
+      const { client, dataMock, headers } = createMock();
 
       const reply = {};
 
-      mock.onPost().reply(config => {
+      dataMock.onPost().reply(config => {
         expect(config.url).toEqual(
-          'https://api.line.me/v2/bot/richmenu/1/content'
+          'https://api-data.line.me/v2/bot/richmenu/1/content'
         );
         expect(config.data).toEqual(undefined);
         expect(config.headers).toEqual(headers);
@@ -688,13 +689,13 @@ describe('Rich Menu', () => {
     it('should call api', async () => {
       expect.assertions(4);
 
-      const { client, mock, headers } = createMock();
+      const { client, dataMock, headers } = createMock();
 
       const reply = Buffer.from('a content buffer');
 
-      mock.onGet().reply(config => {
+      dataMock.onGet().reply(config => {
         expect(config.url).toEqual(
-          'https://api.line.me/v2/bot/richmenu/1/content'
+          'https://api-data.line.me/v2/bot/richmenu/1/content'
         );
         expect(config.data).toEqual(undefined);
         expect(config.headers).toEqual(headers);
@@ -708,15 +709,15 @@ describe('Rich Menu', () => {
 
     it('should work with custom access token', async () => {
       expect.assertions(4);
-      const { client, mock, headers } = createMock({
+      const { client, dataMock, headers } = createMock({
         customAccessToken: CUSTOM_ACCESS_TOKEN,
       });
 
       const reply = Buffer.from('a content buffer');
 
-      mock.onGet().reply(config => {
+      dataMock.onGet().reply(config => {
         expect(config.url).toEqual(
-          'https://api.line.me/v2/bot/richmenu/1/content'
+          'https://api-data.line.me/v2/bot/richmenu/1/content'
         );
         expect(config.data).toEqual(undefined);
         expect(config.headers).toEqual(headers);
@@ -731,9 +732,9 @@ describe('Rich Menu', () => {
     });
 
     it('should return null when no rich menu image found', async () => {
-      const { client, mock } = createMock();
+      const { client, dataMock } = createMock();
 
-      mock.onGet().reply(404, Buffer.from('{"message":"Not found"}'));
+      dataMock.onGet().reply(404, Buffer.from('{"message":"Not found"}'));
 
       const res = await client.downloadRichMenuImage(
         'richmenu-8dfdfc571eca39c0ffcd1f799519c5b5'