+++ CHANGELOG.md
@@ -38,6 +38,20 @@
 - [breaking] refine rich menu getter functions error handling when getting 404
 - [breaking] return null when no user found (#445)
 
+# 0.7.18 / 2020-04-08
+
+- [fix] fix a broken release.
+
+# 0.7.17 / 2020-04-08
+
+### messaging-api-messenger
+
+- [fix] support up to 13 quick replies instead of 11.
+
+### messaging-api-line
+
+- [fix] use data domain to get media and menu content (#557)
+
 # 0.7.16 / 2019-01-29
 
 ### messaging-api-messenger